timeout handing
Added parameters 
search_timeout (takes integer for milliseconds).
search_timeout_msg (takes a string that will display if the swiftype search exceeds the search_timeout value). SwifType was down a few weeks ago due to a DNS attack. Our implementation of the jquery search plugin just hung. This allows for that situation to be dealt with gracefully.